backend/requirements.txt
Flask==3.0.3
Flask-MySQLdb==2.0.0
Flask-CORS==4.0.1
Werkzeug==3.0.3
python-dotenv==1.0.1
pytest==8.2.2
pytest-flask==1.3.0
PyMySQL==1.1.1
cryptography==42.0.8
gunicorn==22.0.0
PyJWT==2.8.0